Add chunking support to ResumableDownloadHandler
Summary:
This adds a chunk_size param to the ResumableDownloadHandler constructor.
This lets us tell boto to download smaller chunks of the file at a time, and should give
the existing retry logic for free.

Test Plan:
In a local script, I access 32 MB chunks at a time. On downloading a
40 MB file, the local file I download has the same md5 as the original file. When
I try downloading a 1 GB file, the Range header for each chunk is as expected.

Reviewers: mattfaus

Reviewed By: mattfaus

Subscribers: sam

Differential Revision: https://phabricator.khanacademy.org/D18709