Fix CertValidatingHTTPSConnection protocol when running under AppEngine
Summary:
AppEngine uses a custom implementation of httplib (imported in
boto/https_connection.py as http_client) that uses urlfetch under the hood. The
custom HTTPConnection and HTTPSConnection classes define a _protocol property
that indicates the protocol to use when constructing URLs for urlfetch.

The boto CertValidatingHTTPSConnection used HTTPConnection as a base class but
didn't define this property. That means CertValidatingHTTPSConnection was
checking the validity of the certificate but then connecting over http (when
running under AppEngine).

This diff changes the base class to HTTPSConnection. Using HTTPSConnection as a
base class is a little awkward because we can't pass `key_file` and `cert_file`
arguments to the parent constructor on AppEngine. The alternative would be to
define a `_protocol` property set to "https", but that ultimately seems more
brittle.

Reviewers: dhruv, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D60550